Combine fetch-with-timeout implementations
There were two competing utility functions for calling fetch with a
timeout. They have been combined into one.
